scripts/requirements.txt
requests
feedparser
psycopg2-binary
pyyaml
numpy
scikit-learn
